+++ Migr8.Test/Basic/SimpleScenarioWithCustomConnectionFactory.cs
@@ -5,12 +5,15 @@ using NUnit.Framework;
 namespace Migr8.Test.Basic
 {
     [TestFixture]
+    [Ignore("Doesn't work - just used to test if connection string could be properly parsed")]
     public class SimpleScenarioWithCustomConnectionFactory : FixtureBase
     {
+        readonly string _connectionString = "Server=(localdb)\\MsSqlLocalDb; Database=migr8_test; Authentication=ManagedIdentity";
+
         [Test]
         public void DoesNothingWhenExecutingEmptyListOfMigrations()
         {
-            Database.Migrate(TestConfig.ConnectionString, new Migrations(Enumerable.Empty<IExecutableSqlMigration>()));
+            Database.Migrate(_connectionString, new Migrations(Enumerable.Empty<IExecutableSqlMigration>()));
         }
 
         [Test]
@@ -23,7 +26,7 @@ namespace Migr8.Test.Basic
                 new TestMigration(3, "test", "CREATE TABLE [Table3] ([Id] int)"),
             };
 
-            Database.Migrate(TestConfig.ConnectionString, new Migrations(migrations));
+            Database.Migrate(_connectionString, new Migrations(migrations));
 
             var tableNames = GetTableNames();
 
@@ -40,11 +43,11 @@ namespace Migr8.Test.Basic
                 new TestMigration(3, "test", "CREATE TABLE [Table3] ([Id] int)"),
             };
 
-            Database.Migrate(TestConfig.ConnectionString, new Migrations(allMigrations.Take(2)));
+            Database.Migrate(_connectionString, new Migrations(allMigrations.Take(2)));
 
             var tableNameAfterFirstTwoMigrations = GetTableNames();
 
-            Database.Migrate(TestConfig.ConnectionString, new Migrations(allMigrations));
+            Database.Migrate(_connectionString, new Migrations(allMigrations));
 
             var tableNameAfterAllMigrations = GetTableNames();