+++ Migr8.Test/FixtureBase.cs
@@ -59,9 +59,7 @@ namespace Migr8.Test
         [TearDown]
         public void TearDown()
         {
-            IDisposable disposable;
-
-            while (_disposables.TryPop(out disposable))
+            while (_disposables.TryPop(out var disposable))
             {
                 Console.WriteLine($"Disposing {disposable}");
                 disposable.Dispose();